requirements.txt
Django==3.2.3
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
flake8-broken-line==0.6.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
isort==5.10.1
pep8-naming==0.13.2
Pillow==9.2.0
python-dotenv==0.21.0
